Move LightTooltip to Components/LightTooltip.tsx
